Split up conversion interface and moved it onto Dataset class
